Rename enums.py to constants.py
