+++ bgstally/constants.py
